Split the scanner by role, changing nothing it does
One file had come to hold four jobs: the data model, every classification
table, the state machine, and the public queries. Each of the day's changes
touched a different job and all of them landed in the same 2,100 lines.
Decided by the director from three options.

The split is by role, not by region. A region has code in two places — its
opening and its continuation arm — and separating those is where the
double-emission defect came from twice; the loop stays whole in mod.rs.

  mod.rs        the model (Piece, EntryToken) and the loop, entire
  tables.rs     every classification table, each with its transcribed source
  regions.rs    what opens each region and where each one ends
  queries.rs    readable_content, readable_for, listing_header_labels
  boundaries.rs balanced_end, delimited_end, is_escaped
  tests.rs      the test module, unchanged

tables.rs is now the one place grammar.md §8's lists correspond to code —
the seam where the external corpus found this week's defects.

Pure moves and import plumbing; no signature, no visibility beyond
pub(crate), no behaviour. Proved rather than claimed: the 992-file corpus
measured with the binary from immediately before the split and immediately
after differs in zero files across every recorded field, the whole suite
passes, and the workspace sweep is clean.
